Fix faulty CMake Regex logic.
CMake was using a cached find of clang-format and clang-tidy, so I didn't know the previous Regex expression I was using did not work.

This fixes it to find a clang-(format|tidy) with one or more version digits. This will NOT work for binaries that do not have the version on the binary name.