Add UI current directory box.
This adds a Gtk::Entry box to display a text box that will show the current directory of the window manager.

A Gtk::Box is used to allow for the widget to be added below the searching file widget.

Finally, this also adds a method to the CurrentDirectoryBar class to modify the displayed text, as the window itself shall be able to modify the directory as well with the navigation bar and directory viewer.