Add DirectoryFilesView base class
This adds the DirectoryFilesView class, and its UI and mock complements, in anticipation for adding the Gtk::ScrollableWindow to display the files in the file system.

This class can handle receiving and forwarding signals when selecting a file or directory in the window. Like other classes, this is also tested to ensure the window properly receives these responses.