Convert GUI tests to use test fixture.
This makes a new WindowTest class for use as a test fixture. This was added to remove as much logic from the GUI test suite, because we shouldn't have to test the tests, or access internals from the window. We need to follow a black-box ideology when testing and not copy or worry about the implementation details of the code itself.