Create FileSystem base class and MockFileSystem for integration with GUI.
This creates the FileSystem base class, which will be used to abstract away the exact underlying file system from the GUI, and facililate creating different file system support, such as network and POSIX file systems.

This also creates the MockFileSystem class to be used to test the GUI's interactions with a file system. This includes tests for the MockFileSystem itself, since I wasn't able to find an one online.