Fix MockFileManager handling of empty string.
The MockFileManager::GetDirectoryFiles() fails to hande the edge case of an empty directory given. This should have been tested for, but was overlooked. absl::StrSplit() will still insert at least one string even when the input is empty. This instead directly checks for an empty string to handle the case of an empty path entered.