Show GoogleTest how to print Glib::ustring.
This overloads the PrintTo() method in GoogleTest to allow printing of Glib::ustring types with my own implementation. This was chosen over an operator<< overload because GoogleTest was flat out ignoring that implementation. Has to be in the Glib namespace because of C++ name lookup rules requiring this to be in the same namepsace as the ustring type to find it using argument dependent lookup.