Update GUI text suite for directory bar API update.
A previous commit updates CurrentDirectoryBar::SetDisplayedDirectory() to not return a bool because there isn't many ways to signify an error from GTKMM. Any error-checking and input validation will be done by the window itself. CurrentDirectoryBar::SetDisplayedDirectory() method will simply assign to its internal data structures.

The GUI test suite was not updated to account for this API update, so updates any test setup to account for this as well.