Remove tests for Internal window updates.
The GUI test had some bad tests that were testing an internal method for the window. In accordance to testing via the public API and black-box testing, these tests are not neccessary. We should only test via the way a user will actually rely on this method, which is through button presses and a directory entered via the current directory bar widget.

This simply removes the tests that tested the Window::UpdateDirectory API from the GUI test suite, since this API should not be exposed to the public.