Change directory chaning mechanism.
This updates the internal directory mechanism since the previous one was faulty.

The previous one had a directory changing mechanism where each change (go back, go forward, and go up) would call UpdateDirectory() at the end. UpdateDirectory() would validate and update the history. However, this shouldn't be the case since different windows have different ways to update the history. The implementation for this was even broken because some of the method didn't call UpdateDirectory().

In this update, the same control flow goes but now there is a method for absolute directory changes. The absolute directory change method validates the input. UpdateDirectory() is then called on each mechanism. UpdateDirectory() does not validate the input, so it expects valid input placed through it. The go back and go forward mechanisms use the history to get their directories, and it assumes that the history data will always be valid. Going up and absolute directory changes are checked for validity. Each directory change mechanism handles the history in its own way.

There is also the addition of the Window::RefreshWindowComponents() virtual method, which allow any type of window (Mock or UI), to refresh the window internal widgets. This was necessary because of GTKMM requiring a call to show_all() to display all the updated widgets if their data was manipulated.