Add API to modify files shown on directory view.
This adds the AddFile and RemoveAllFiles methods to the DirectoryFilesView class. This will give the window an interface to update the files shown on the DirectoryFilesView's internal widgets. The DirectoryFileView forces any clases to override this method to indicate how it will add files to its internal widgets (great for mocking because we can add mock files to a mock file list).