Implement UIDirectoryFilesView with GTKMM widgets.
This implements the AddFile, RemoveAllFiles and OnFileClick callbacks with GTK::ToggleButton for the UIDirectoryFilesView class, so they can add and remove widgets from the GTK::ScrolledWindow, which contains the files from the file system.

Note this does not differentiate files from directories just yet. The window treats them both as directories, which means files get navigated to. We need to diversify the FIleSystem::GetDirectoryFiles() API to signify when we have a file or a directory, so we can add appropriate logic for the two.