Switch back and forward buttons. Closes #1
The UINavBar was inserting the navigation buttons into the Gtk::Box container in the wrong order, causing the up button showing up first, then the forward and back button. This switches the placement to resolve this issue to have the back button, forward button, and up button in that order.