Add file abstraction for different file systems.
This converts the program to using the ::File class to add more information about a file we retrieve from any supported file system. This improves upon only using a string for the file name, which gave us no way to determine other properties of the file. The ::File class can be used to add futher metadata about the file as needed.