Add icons to discern folders and regular files.
This makes use of the GTKMM to load icons of a folder and a regular file to differentiate between directories and regular files. The icons are currently loaded multiple times for all the files, and freed when a directory is changed. This can be optimized by loading the icons only once in the future. The file type is determined by the File object passed in.

This also adjusts the DirectoryFileView class to support the AddFile API change for File objects instead of strings only.