Implement network connection interface.
This adds the NetworkInterface API that will wrap any future system networking APIs. This also adds the NetworkConnection API that can be used to establish an active connection and perform duplex communication to a networking endpoint. It uses the NetworkInterface API to do this in a platform-agnostic manner. The NetworkInterface class can be derived from to implement the actual API calls to the system.

The MockNetworkInterface class was created to test the NetworkConnection class. The MockNetworkInterface class can be used to work with mock sockets, send back a certain amount of bytes, and to accept and send bytes in random intervals. It might be worthwile investing in making the MockNetworkInterface have a declarative programming model to make the behavior more customizable.