Revise network interface due to double free of LL.
This changes the NetworkAddressInfo structure to be a container of an array of nodes containing network information, rather than representing an individual network address node. This was done because we were previously deleting each node in the list that was retrieved from the POSIX linked list method. In that method, we should only free the head of the linked list. We were deleting each node in the linked list instead, which resulted in double frees.

The NetworkAddressInfoNode class now represents each node of all the possible addresses for a network endpoint. The NetworkAddressInfo contains a list of these nodes. The NetworkAddressInfoNode class can be extended for all future interfaces that this program can adhere to.

This also changes all the implementations that used the old structure to adhere to the new interface.